+++ src/react/components/button/stories/variations.js
@@ -3,116 +3,77 @@ import { desktopPlus320 } from '../../../stories/helpers/viewports';
 
 const privateConsts = require('../private/constants');
 
-const storiesWithKnobsKindName = 'Components/XUIButton';
-const storiesWithVariationsKindName = `${storiesWithKnobsKindName}/Tests`;
+export const buttonStoriesWithKnobsKindName = 'Components/XUIButton';
+export const buttonStoriesWithVariationsKindName = `${buttonStoriesWithKnobsKindName}/Tests`;
+
+export const iconButtonStoriesWithKnobsKindName = 'Components/XUIIconButton';
+export const iconButtonStoriesWithVariationsKindName = `${iconButtonStoriesWithKnobsKindName}/Tests`;
+
+export const buttonGroupStoriesWithKnobsKindName = 'Components/XUIButtonGroup';
+export const buttonGroupStoriesWithVariationsKindName = `${buttonGroupStoriesWithKnobsKindName}/Tests`;
+
+export const splitButtonGroupStoriesWithKnobsKindName = 'Components/XUISplitButtonGroup';
+export const splitButtonGroupStoriesWithVariationsKindName = `${splitButtonGroupStoriesWithKnobsKindName}/Tests`;
+
 const sizes = Object.keys(privateConsts.sizeClassNames);
 const iconSizes = Object.keys(privateConsts.iconSizeClassNames);
-const variations = [
+
+const buttonVariations = [
   {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: 'as Disabled',
     isDisabled: true,
     value: 'Disabled button',
   },
   {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: 'with multiline content',
     viewports: desktopPlus320,
     value:
       'Some very long text to test how buttons behave when their content is longer than the alotted space. Some very long text to test how buttons behave when their content is longer than the alotted space.',
   },
   {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: 'with caret',
     contentsKey: 'withCaret',
     hasCaret: true,
     variant: 'create',
   },
   {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: 'small with caret',
     contentsKey: 'withCaret',
     hasCaret: true,
     size: 'small',
   },
   {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: 'with left icon',
     contentsKey: 'withIcon',
     leftIcon: settings,
   },
   {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: 'xsmall with left icon',
     contentsKey: 'withIcon',
     leftIcon: settings,
     size: 'xsmall',
   },
   {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: 'with right icon',
     contentsKey: 'withIcon',
     rightIcon: settings,
   },
   {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: 'xsmall with right icon',
     contentsKey: 'withIcon',
     rightIcon: settings,
     size: 'xsmall',
   },
   {
-    storyKind: storiesWithVariationsKindName,
-    storyTitle: 'as a Group',
-    contentsKey: 'asGroup',
-    componentType: 'XUIButtonGroup',
-  },
-  {
-    storyKind: storiesWithVariationsKindName,
-    storyTitle: 'as a Group with multiline content',
-    viewports: desktopPlus320,
-    contentsKey: 'asMultiGroup',
-    componentType: 'XUIButtonGroup',
-  },
-  {
-    storyKind: storiesWithVariationsKindName,
-    storyTitle: 'as a small Group',
-    contentsKey: 'asGroup',
-    componentType: 'XUIButtonGroup',
-    size: 'small',
-  },
-  {
-    storyKind: storiesWithVariationsKindName,
-    storyTitle: 'as a Split Button Group',
-    contentsKey: 'asSplitGroup',
-    componentType: 'XUISplitButtonGroup',
-    variant: 'main',
-  },
-  {
-    storyKind: storiesWithVariationsKindName,
-    storyTitle: 'as a Split Button Group with multi-line text',
-    componentType: 'XUISplitButtonGroup',
-    variant: 'main',
-    contentsKey: 'asSplitGroupMulti',
-  },
-  {
-    storyKind: storiesWithVariationsKindName,
-    storyTitle: 'as a small negative Split Button Group',
-    contentsKey: 'asSplitGroup',
-    componentType: 'XUISplitButtonGroup',
-    variant: 'negative',
-    size: 'small',
-  },
-  {
-    storyKind: storiesWithVariationsKindName,
-    storyTitle: 'as an xsmall Split Button Group with a dropdown',
-    contentsKey: 'asSplitGroupDropdown',
-    componentType: 'XUISplitButtonGroup',
-    size: 'xsmall',
-    variant: 'main',
-  },
-  {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: 'as small fullWidth',
     viewports: desktopPlus320,
     value: 'fullWidth always',
@@ -120,7 +81,7 @@ const variations = [
     size: 'small',
   },
   {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: 'as large fullWidth with an icon',
     contentsKey: 'withIcon',
     leftIcon: settings,
@@ -129,17 +90,73 @@ const variations = [
     fullWidth: 'always',
   },
   {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: 'as medium fullWidth small-down',
     viewports: desktopPlus320,
     value: 'fullWidth small-down',
     fullWidth: 'small-down',
   },
   {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: `with short content and hasMinLoaderWidth`,
     value: 'B',
     hasMinLoaderWidth: true,
+  },
+];
+
+const buttonGroupVariations = [
+  {
+    storyKind: buttonGroupStoriesWithVariationsKindName,
+    storyTitle: 'as a Group',
+    contentsKey: 'asGroup',
+    componentType: 'XUIButtonGroup',
+  },
+  {
+    storyKind: buttonGroupStoriesWithVariationsKindName,
+    storyTitle: 'as a Group with multiline content',
+    viewports: desktopPlus320,
+    contentsKey: 'asMultiGroup',
+    componentType: 'XUIButtonGroup',
+  },
+  {
+    storyKind: buttonGroupStoriesWithVariationsKindName,
+    storyTitle: 'as a small Group',
+    contentsKey: 'asGroup',
+    componentType: 'XUIButtonGroup',
+    size: 'small',
+  },
+];
+
+const splitButtonGroupVariations = [
+  {
+    storyKind: splitButtonGroupStoriesWithVariationsKindName,
+    storyTitle: 'as a Split Button Group',
+    contentsKey: 'asSplitGroup',
+    componentType: 'XUISplitButtonGroup',
+    variant: 'main',
+  },
+  {
+    storyKind: splitButtonGroupStoriesWithVariationsKindName,
+    storyTitle: 'as a Split Button Group with multi-line text',
+    componentType: 'XUISplitButtonGroup',
+    variant: 'main',
+    contentsKey: 'asSplitGroupMulti',
+  },
+  {
+    storyKind: splitButtonGroupStoriesWithVariationsKindName,
+    storyTitle: 'as a small negative Split Button Group',
+    contentsKey: 'asSplitGroup',
+    componentType: 'XUISplitButtonGroup',
+    variant: 'negative',
+    size: 'small',
+  },
+  {
+    storyKind: splitButtonGroupStoriesWithVariationsKindName,
+    storyTitle: 'as an xsmall Split Button Group with a dropdown',
+    contentsKey: 'asSplitGroupDropdown',
+    componentType: 'XUISplitButtonGroup',
+    size: 'xsmall',
+    variant: 'main',
   },
 ];
 
@@ -147,8 +164,8 @@ Object.keys(privateConsts.textButtonVariants).forEach((buttonVariant, index) => 
   const sizesToIterate = [sizes[index % sizes.length]];
 
   sizesToIterate.forEach(size =>
-    variations.push({
-      storyKind: storiesWithVariationsKindName,
+    buttonVariations.push({
+      storyKind: buttonStoriesWithVariationsKindName,
       storyTitle: `as ${size} ${buttonVariant}`,
       href: buttonVariant === 'link' ? '#' : undefined,
       isLink: buttonVariant === 'link',
@@ -160,8 +177,8 @@ Object.keys(privateConsts.textButtonVariants).forEach((buttonVariant, index) => 
 });
 
 sizes.forEach(sizeVariant => {
-  variations.push({
-    storyKind: storiesWithVariationsKindName,
+  buttonVariations.push({
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: `${sizeVariant} with a Loader`,
     size: sizeVariant,
     isLoading: true,
@@ -169,27 +186,41 @@ sizes.forEach(sizeVariant => {
     className: 'xui-loader-static', // Prevent dots from animating and causing diffs due to timing issues
     value: 'This is a button',
   });
-  variations.push({
-    storyKind: storiesWithVariationsKindName,
+  buttonVariations.push({
+    storyKind: buttonStoriesWithVariationsKindName,
     storyTitle: `${sizeVariant} with short content`,
     size: sizeVariant,
     value: 'B',
   });
 });
 
+const iconButtonVariations = [];
+
 iconSizes.forEach(iconSize => {
   const iconButton = {
-    storyKind: storiesWithVariationsKindName,
+    storyKind: iconButtonStoriesWithVariationsKindName,
     storyTitle: `${iconSize} iconButton`,
     componentType: 'XUIIconButton',
     size: iconSize,
   };
-  variations.push(iconButton);
-  variations.push({
+  iconButtonVariations.push(iconButton);
+  iconButtonVariations.push({
     ...iconButton,
     isInverted: true,
     storyTitle: `${iconSize} inverted ${iconButton.storyTitle}`,
   });
 });
 
-export { storiesWithVariationsKindName, storiesWithKnobsKindName, variations };
+export {
+  buttonVariations,
+  iconButtonVariations,
+  buttonGroupVariations,
+  splitButtonGroupVariations,
+};
+
+export const variations = [
+  ...buttonVariations,
+  ...iconButtonVariations,
+  ...buttonGroupVariations,
+  ...splitButtonGroupVariations,
+];